BARK-113: Move toUserPii and toEncryptedUserPii into user-mappers.ts
