+++ src/lib/bark/mappers/user-mappers.ts
@@ -1,5 +1,15 @@
 import { BarkContext } from "../bark-context";
-import { UserPii } from "../models/user-pii";
+import { UserPii, UserPiiSchema } from "../models/user-pii";
+
+export async function toUserPii(
+  context: BarkContext,
+  userEncryptedPii: string,
+): Promise<UserPii> {
+  const { piiEncryptionService } = context;
+  const jsonEncoded =
+    await piiEncryptionService.getDecryptedData(userEncryptedPii);
+  return UserPiiSchema.parse(JSON.parse(jsonEncoded));
+}
 
 export async function toEncryptedUserPii(
   context: BarkContext,
